moved equivalence class from hypopplactic monoid to WordMonoid; removed duplicated TestSuite
